Initial band component to change peak parameters of filter

+++ plugin/include/SimpleParametricEq/filters/BiquadFilter.h
@@ -98,10 +98,10 @@ public:
         }
 
         const auto ω = 2.0 * juce::MathConstants<double>::pi * freq / sampleRate_;
-        const auto cos1 = std::cos(ω);
-        const auto sin1 = std::sin(ω);
-        const auto cos2 = std::cos(2.0 * ω);
-        const auto sin2 = std::sin(2.0 * ω);
+        const auto cos1 = static_cast<float>(std::cos(ω));
+        const auto sin1 = static_cast<float>(std::sin(ω));
+        const auto cos2 = static_cast<float>(std::cos(2.0 * ω));
+        const auto sin2 = static_cast<float>(std::sin(2.0 * ω));
 
         const auto numReal = b0_ + b1_ * cos1 + b2_ * cos2;
         const auto numImag = -b1_ * sin1 - b2_ * sin2;
@@ -112,7 +112,7 @@ public:
         const auto numMag2 = numReal * numReal + numImag * numImag;
         const auto denMag2 = denReal * denReal + denImag * denImag;
 
-        if (denMag2 <= 0.0) {
+        if (denMag2 <= 0.0f) {
             return 1.0f;
         }
